+++ Firmware/Firmware Build Files/pysquared_rp2040_flight_controller_v1a/pico-sdk-configboard.h
@@ -1,0 +1,4 @@
+// Put board-specific pico-sdk definitions here. This file must exist.
+
+// Allow extra time for xosc to start.
+#define PICO_XOSC_STARTUP_DELAY_MULTIPLIER 64
