--- Firmware/pysquared-clockspeed/mpconfigboard.h
@@ -1,28 +1,0 @@
-// This file is part of the CircuitPython project: https://circuitpython.org
-//
-// SPDX-FileCopyrightText: Copyright (c) 2021 Scott Shawcroft for Adafruit Industries
-//
-// SPDX-License-Identifier: MIT
-
-#pragma once
-
-#define MICROPY_HW_BOARD_NAME "PySquaredFCv4"
-#define MICROPY_HW_MCU_NAME "rp2040"
-
-#define MICROPY_HW_NEOPIXEL (&pin_GPIO24)
-
-#define CIRCUITPY_INTERNAL_NVM_SIZE (4 * 1024)
-#define CIRCUITPY_INTERNAL_FLASH_FILESYSTEM_SIZE (48 * 1024)
-#define BOARD_FLASH_SIZE (FLASH_SIZE - 0x4000 - CIRCUITPY_INTERNAL_NVM_SIZE)
-
-#define CIRCUITPY_DRIVE_LABEL "PYSQUARED"
-
-#define DEFAULT_I2C_BUS_SCL (&pin_GPIO3)
-#define DEFAULT_I2C_BUS_SDA (&pin_GPIO2)
-
-#define DEFAULT_SPI_BUS_SCK (&pin_GPIO10)
-#define DEFAULT_SPI_BUS_MOSI (&pin_GPIO8)
-#define DEFAULT_SPI_BUS_MISO (&pin_GPIO11)
-
-#define DEFAULT_UART_BUS_RX (&pin_GPIO1)
-#define DEFAULT_UART_BUS_TX (&pin_GPIO0)
