--- Firmware/pysquared-clockspeed/mpconfigboard.mk
@@ -1,11 +1,0 @@
-USB_VID = 0x1209
-USB_PID = 0x0011
-USB_PRODUCT = "FLIGHT_CONTROLLER"
-USB_MANUFACTURER = "BRONCO_SPACE"
-
-CHIP_VARIANT = RP2040
-CHIP_FAMILY = rp2
-
-EXTERNAL_FLASH_DEVICES = "GD25Q64C,W25Q64JVxQ"
-
-CFLAGS += -DSET_CLOCK_SPEED
