--- Firmware/pysquared-clockspeed/pico-sdk-configboard.h
@@ -1,4 +1,0 @@
-// Put board-specific pico-sdk definitions here. This file must exist.
-
-// Allow extra time for xosc to start.
-#define PICO_XOSC_STARTUP_DELAY_MULTIPLIER 64
